Fix compatibility issue with AnimeJS v3.x
Prior to AnimeJS v3.x, animation timelines were guaranteed to make
progress. This is no longer the case; timelines are now suspended if
the browser tab is inactive. This throws the number of active phrases
off balance, because destruction of a phrase depends on its timeline
having finished. This led to a disturbing number of active phrases at
the same time if the tab is left inactive for too long.

To work around that issue, introduce a maximum number of active phrases
and have the phrase engine enforce that limit.